Fix shell bangs in prep section
The dbixs_rev.pl should not be installed probably. It looks like
an upstream mistake because EU::MM installs everything with .pl
extension.

The correct shell bang in dbixs_rev.pl triggers rpmlint error, but the
correct bang is needed becuse it is run during %build phase from the
Makefile.